+++ kivy/tests/conftest.py
@@ -1,22 +1,13 @@
 import pytest
-from functools import wraps
 import gc
 import asyncio
 import time
 import os.path
-from kivy.graphics.cgl import cgl_get_backend_name
+from kivy.tests import async_sleep
 
 try:
-    import pytest_asyncio
-    async_sleep = asyncio.sleep
+    from pytest_trio import trio_fixture
 except ImportError:
-    try:
-        from pytest_trio import trio_fixture
-        import trio
-        async_sleep = trio.sleep
-    except ImportError:
-        pass
-else:
     @pytest.fixture()
     def nursery():
         pass
@@ -36,43 +27,6 @@ def pytest_runtest_setup(item):
         previousfailed = getattr(item.parent, "_previousfailed", None)
         if previousfailed is not None:
             pytest.xfail("previous test failed (%s)" % previousfailed.name)
-
-
-def async_run(func=None, app_cls_func=None):
-    def inner_func(func):
-        if 'mock' == cgl_get_backend_name():
-            return pytest.mark.skip(
-                reason='Skipping because gl backend is set to mock')(func)
-
-        try:
-            import kivy.tests.async_common
-        except SyntaxError:
-            return pytest.mark.skip(
-                reason='Skipping because graphics tests are not supported on '
-                       'py3.5, only on py3.6+')(func)
-
-        if app_cls_func is not None:
-            func = pytest.mark.parametrize(
-                "kivy_app", [[app_cls_func], ], indirect=True)(func)
-
-        try:
-            import pytest_asyncio
-            return pytest.mark.asyncio(func)
-        except ImportError:
-            try:
-                import trio
-                from pytest_trio import trio_fixture
-                func._force_trio_fixture = True
-                return func
-            except ImportError:
-                return pytest.mark.skip(
-                    reason='Either pytest_asyncio or pytest_trio must be '
-                    'installed to run asyncio tests')(func)
-
-    if func is None:
-        return inner_func
-
-    return inner_func(func)
 
 
 @pytest.fixture()
